requirements.txt
aiogram==3.13.1
openai==1.54.3
python-dotenv==1.0.1
aiohttp==3.10.5
aiofiles==24.1.0
httpx==0.27.2
aiosqlite==0.20.0
pydantic==2.9.2
PyYAML==6.0.2
youtube-transcript-api==0.6.2
markdownify==0.13.1
schedule
Jinja2==3.1.4
whisperx
torch
torchaudio

